Split astra-plan module boundaries
