+++ rust/crates/astra-plan/src/lib.rs
@@ -1,11 +1,19 @@
-//! Astra Plan — goal decomposition engine for breaking complex tasks into subtasks.
+//! Astra Plan — shared plan state, execution helpers, and persistence boundaries.
 
-pub mod decompose;
+pub mod execution;
 pub mod repository;
+pub mod resume;
+pub mod state;
 
-pub use decompose::*;
+pub use execution::{
+    FileConflict, ParallelGroups, analyze_parallelism, format_subtask_prompt_with_operator_notes,
+    subtask_requires_browser_verification,
+};
 pub use repository::{
     CloudPlanRepository, InMemoryPlanRepository, NewStepRun, PlanListFilter, PlanLoadError,
-    PlanRepository, PlanStepRun, SavedPlanInfo, plan_resume_digest, plan_resume_hint_for_session,
-    plan_resume_prompt_hint,
+    PlanRepository, PlanStepRun, SavedPlanInfo,
+};
+pub use resume::{plan_resume_digest, plan_resume_hint_for_session, plan_resume_prompt_hint};
+pub use state::{
+    ExecutionTimeline, PlanExecutionConfig, PlanModeState, TimelineEvent, TimelineEventKind,
 };
